+++ common/test/python/curator/test_scheduling.py
@@ -3,7 +3,7 @@
 from datetime import date
 from typing import Optional
 
-from curator.scheduling import FileModel
+from curator.scheduling_models import FileModel
 
 
 def fm(
@@ -42,6 +42,7 @@ class TestFileModel:
         assert fm("NACCXXX_NP-RECORD-2012-02-10_NP.json").visit_pass == "pass1"
         assert fm("NACCXXX_MDS-RECORD-2006-03-23_MDS.json").visit_pass == "pass1"
         assert fm("NACCXXX_MILESTONE-2011-10-26_MLST.json").visit_pass == "pass1"
+        assert fm("NACCXXX_FORMS-VISIT-2_MEDS.json").visit_pass == "pass1"
 
         assert fm("NACCXXX_apoe_genotype.json").visit_pass == "pass1"
         assert fm("NACCXXX_niagads_availability.json").visit_pass == "pass1"
@@ -69,7 +70,6 @@ class TestFileModel:
 
         # pass0
         assert fm("NACCXXX_FORMS-VISIT-5_UDS.json").visit_pass == "pass0"
-        assert fm("NACCXXX_FORMS-VISIT-2_MEDS.json").visit_pass == "pass0"
 
         # invalid/unknown pass
         assert fm("invalid.json").visit_pass is None