+++ gear/regression_curator/src/python/regression_curator_app/run.py
@@ -36,10 +36,10 @@ class RegressionCuratorVisitor(GearExecutionEnvironment):
         client: ClientWrapper,
         project: ProjectAdaptor,
         s3_qaf_file: str,
-        s3_mqt_file: str,
         keep_fields: List[str],
         filename_pattern: str,
         error_outfile: str,
+        s3_mqt_file: Optional[str] = None,
         blacklist_file: Optional[InputFileWrapper] = None,
     ):
         super().__init__(client=client)
@@ -73,9 +73,8 @@ class RegressionCuratorVisitor(GearExecutionEnvironment):
         s3_qaf_file = context.config.get("s3_qaf_file", None)
         s3_mqt_file = context.config.get("s3_mqt_file", None)
 
-        for file in [s3_qaf_file, s3_mqt_file]:
-            if not file:
-                raise GearExecutionError("Required baseline file (QAF/MQT) missing")
+        if not s3_qaf_file:
+            raise GearExecutionError("QAF file missing")
 
         keep_fields = parse_string_to_list(context.config.get("keep_fields", ""))
         filename_pattern = context.config.get("filename_pattern", "*UDS.json")
@@ -89,6 +88,9 @@ class RegressionCuratorVisitor(GearExecutionEnvironment):
         blacklist_file = InputFileWrapper.create(
             input_name="blacklist_file", context=context
         )
+
+        if context.config.get("debug", False):
+            logging.basicConfig(level=logging.DEBUG)
 
         return RegressionCuratorVisitor(
             client=client,