+++ docs/regression_curator/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this gear are documented in this file.
 
+## 0.1.2
+
+* Adds `debug` argument to reduce verbosity of logging statements
+* Updates MQT regression testing to not be required
+
 ## 0.1.1
 
 * Changes type to `analysis`